+++ CHANGELOG.md
@@ -2,6 +2,7 @@
 
 #### _***All Apps and Devices***_
 - [FIX] Bugfixes and Optimizations.
+- [FIX] Fixed Device Info Page and Added Test Buttons
 
 ### _**Release (v4.1.9.2) - [July 22nd, 2021]**_